requirements.txt
# iac
aws-cdk.core
aws-cdk.aws-iam
aws-cdk.aws-s3
aws-cdk.aws-lambda
aws-cdk.aws-apigateway
aws-cdk.aws-ec2
aws-cdk.aws-rds
aws_cdk.aws_elasticloadbalancingv2
aws_cdk.aws_autoscaling

# lambda
boto3
marshmallow-jsonapi
flask-smorest

# testing
pytest
flake8
